Add files via upload
HTML and CSS files now have an identifier ID titled "align_middle" that will be used to position certain elements to the middle of the webpage. The div elements for this have not been fully added and more will be required. This pertains to issue #1. Also, multiple comments have been added as instructions for contributors.